# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: ASI:BUILD API
  description: |
    Comprehensive API for the ASI:BUILD Superintelligence Framework.
    
    The ASI:BUILD framework provides access to 47 integrated subsystems including:
    - Consciousness Engine with multi-layered awareness
    - Quantum Computing with hybrid ML processing
    - Reality Manipulation with physics control
    - Divine Mathematics for transcendent computation
    - Absolute Infinity capabilities
    - Swarm Intelligence coordination
    - Bio-inspired systems and neuromorphic processing
    
    ## Security
    All endpoints require JWT Bearer token authentication with role-based access control.
    
    ## Safety Protocols
    The API implements comprehensive safety protocols including reality locks,
    consciousness protection, god mode controls, and emergency shutdown capabilities.
    
    ## Rate Limiting
    Rate limits are enforced per IP address with varying limits based on endpoint sensitivity.
  version: 1.0.0
  contact:
    name: ASI:BUILD API Support
    email: api-support@asi-build.ai
    url: https://asi-build.ai/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.asi-build.ai
    description: Production server
  - url: https://sandbox.asi-build.ai
    description: Sandbox server for testing
security:
  - bearerAuth: []
...
paths:
...
  /api/status:
    get:
      tags:
        - Core API
      summary: Get system status
      description: Returns comprehensive system status information
      responses:
        '200':
          description: System status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemStatus'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /api/emergency/shutdown:
    post:
      tags:
        - Emergency
      summary: Emergency shutdown
      description: Initiate emergency shutdown of ASI:BUILD system (requires admin role)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reason
              properties:
                reason:
                  type: string
                  description: Reason for emergency shutdown
                  minLength: 10
                  maxLength: 500
      responses:
        '200':
          description: Emergency shutdown initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmergencyResponse'
...
  /reality/manipulate:
    post:
      tags:
        - Reality Engine
      summary: Reality manipulation (God Mode Required)
      description: Manipulate reality parameters (requires god mode)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RealityManipulateRequest'
      responses:
        '200':
          description: Reality manipulation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RealityManipulateResponse'
        '403':
          $ref: '#/components/responses/RealityLocked'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SystemStatus:
      type: object
      properties:
        state:
          type: string
          enum: [offline, initializing, active, god_mode, emergency_mode, shutdown]
        uptime:
          type: number
        active_subsystems:
          type: integer
        safety_level:
          type: string
        reality_locked:
          type: boolean
        god_mode_enabled:
          type: boolean
        human_oversight_active:
          type: boolean
        system_metrics:
          type: object
          properties:
            memory_usage:
              type: number
            cpu_usage:
              type: number
            disk_usage:
              type: number
            active_connections:
              type: integer
...
    EmergencyResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        initiated_by:
          type: string
        reason:
          type: string
        timestamp:
          type: number
...
    RealityManipulateRequest:
      type: object
      required:
        - target
        - operation
      properties:
        target:
          type: string
          enum: [spacetime, matter, energy, information]
        operation:
          type: string
          enum: [create, modify, destroy, transform]
        parameters:
          type: object
        safety_override:
          type: string
          description: Emergency safety override code
    RealityManipulateResponse:
      type: object
      properties:
        success:
          type: boolean
        manipulation_id:
          type: string
        effect:
          type: object
        safety_warnings:
          type: array
          items:
            type: string
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string
        error_code:
          type: string
        timestamp:
          type: number
        trace_id:
          type: string
  responses:
...
    ServiceUnavailable:
      description: Service unavailable - System not ready
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RealityLocked:
      description: Reality manipulation not permitted - Reality locks active
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'