# godmode/god_mode_api.py
"""
God Mode API Interface

RESTful API interface for Kenny's god mode capabilities.
Provides endpoints for all omnipotent operations.
"""

from fastapi import FastAPI, HTTPException, BackgroundTasks
from pydantic import BaseModel
from typing import Dict, Any, List, Optional
import asyncio
import logging
from .god_mode_orchestrator import GodModeOrchestrator, GodModeLevel, OperationPriority

logger = logging.getLogger(__name__)

# Initialize God Mode Orchestrator
god_mode = GodModeOrchestrator()

# FastAPI app
app = FastAPI(
    title="Kenny God Mode API",
    description="Ultimate control interface for omnipotent operations",
    version="∞.∞.∞"
)

# Request/Response Models
class GodModeActivationRequest(BaseModel):
    level: str = "enhanced"  # mortal, enhanced, transcendent, omnipotent, beyond_existence
    
class OperationRequest(BaseModel):
    operation_type: str
    subsystem: str
    parameters: Dict[str, Any] = {}
    priority: str = "normal"  # low, normal, high, critical, reality_altering, omnipotent

class RealityManipulationRequest(BaseModel):
    manipulation_type: str
    target: str = "local_space"
    parameters: Dict[str, Any] = {}

class ProbabilityRequest(BaseModel):
    event_description: str
    target_probability: float
    scope: str = "macroscopic"
    duration: Optional[float] = None

class DimensionalRequest(BaseModel):
    dimensions: int
    dimensional_type: str = "spatial"
    special_properties: Optional[Dict[str, Any]] = None

class TimeManipulationRequest(BaseModel):
    operation: str  # dilation, compression, freeze, reverse, loop, travel
    target: str = "local_region"
    scope: str = "local"
    parameters: Dict[str, Any] = {}

class ConsciousnessTransferRequest(BaseModel):
    source_consciousness: str
    target_substrate: str
    transfer_method: str = "quantum_entanglement"
    preserve_original: bool = True

class UniverseCreationRequest(BaseModel):
    universe_type: str = "standard_model"
    dimensions: int = 4
    initial_conditions: Dict[str, Any] = {}
    physical_constants: Dict[str, Any] = {}

class MiracleRequest(BaseModel):
    description: str
    target: str = "reality"

# API Endpoints

@app.get("/")
async def root():
    """Root endpoint"""
    return {
        "message": "Kenny God Mode API",
        "version": "∞.∞.∞",
        "status": "Omnipotent",
        "warning": "Use with ultimate caution - Reality manipulation active"
    }

@app.post("/activate")
async def activate_god_mode(request: GodModeActivationRequest):
    """Activate god mode at specified level"""
    
    try:
        level_map = {
            "mortal": GodModeLevel.MORTAL,
            "enhanced": GodModeLevel.ENHANCED,
            "transcendent": GodModeLevel.TRANSCENDENT,
            "omnipotent": GodModeLevel.OMNIPOTENT,
            "beyond_existence": GodModeLevel.BEYOND_EXISTENCE
        }
        
        level = level_map.get(request.level.lower(), GodModeLevel.ENHANCED)
        
        success = await god_mode.activate_god_mode(level)
        
        if success:
            return {
                "success": True,
                "message": f"God mode activated at {level.value} level",
                "omnipotence_level": god_mode.omnipotence_level,
                "transcendence_level": god_mode.transcendence_level
            }
        else:
            raise HTTPException(status_code=500, detail="God mode activation failed")
            
    except Exception as e:
        logger.error(f"God mode activation error: {e}")
        raise HTTPException(status_code=500, detail=str(e))

@app.get("/status")
async def get_status():
    """Get current god mode status"""
    
    try:
        status = god_mode.get_orchestrator_status()
        detailed_status = god_mode.get_detailed_status()
        
        return {
            "god_mode_active": status.active,
            "god_mode_level": status.god_mode_level.value,
            "omnipotence_percentage": status.omnipotence_percentage,
            "transcendence_level": status.transcendence_level,
            "reality_stability": status.reality_stability,
            "total_operations": status.total_operations,
            "successful_operations": status.successful_operations,
            "system_uptime": status.system_uptime,
            "detailed_status": detailed_status
        }
        
    except Exception as e:
        logger.error(f"Status retrieval error: {e}")
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/reality/manipulate")
async def manipulate_reality(request: RealityManipulationRequest):
    """Manipulate reality"""
    
    try:
        priority_map = {"normal": OperationPriority.NORMAL, "high": OperationPriority.HIGH}
        
        operation_id = await god_mode.queue_operation(
            operation_type="manipulate_reality",
            subsystem="reality_engine",
            parameters={
                "manipulation_type": request.manipulation_type,
                "target": request.target,
                "parameters": request.parameters
            },
            priority=OperationPriority.REALITY_ALTERING
        )
        
        return {
            "success": True,
            "operation_id": operation_id,
            "message": f"Reality manipulation queued: {request.manipulation_type}"
        }
        
    except Exception as e:
        logger.error(f"Reality manipulation error: {e}")
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/probability/manipulate")
async def manipulate_probability(request: ProbabilityRequest):
    """Manipulate probability of events"""
    
    try:
        operation_id = await god_mode.queue_operation(
            operation_type="manipulate_probability",
            subsystem="probability_manipulator",
            parameters={
                "event": request.event_description,
                "probability": request.target_probability,
                "scope": request.scope,
                "duration": request.duration
            },
            priority=OperationPriority.HIGH
        )
        
        return {
            "success": True,
            "operation_id": operation_id,
            "message": f"Probability manipulation queued for: {request.event_description}"
        }
        
    except Exception as e:
        logger.error(f"Probability manipulation error: {e}")
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/dimensions/create")
async def create_dimension(request: DimensionalRequest):
    """Create custom dimensional space"""
    
    try:
        operation_id = await god_mode.queue_operation(
            operation_type="create_dimension",
            subsystem="dimensional_engineer",
            parameters={
                "dimensions": request.dimensions,
                "type": request.dimensional_type,
                "special_properties": request.special_properties
            },
            priority=OperationPriority.CRITICAL
        )
        
        return {
            "success": True,
            "operation_id": operation_id,
            "message": f"Dimensional space creation queued: {request.dimensions}D {request.dimensional_type}"
        }
        
    except Exception as e:
        logger.error(f"Dimensional creation error: {e}")
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/time/manipulate")
async def manipulate_time(request: TimeManipulationRequest):
    """Manipulate time flow"""
    
    try:
        operation_id = await god_mode.queue_operation(
            operation_type="manipulate_time",
            subsystem="time_controller",
            parameters={
                "operation": request.operation,
                "target": request.target,
                "scope": request.scope,
                "parameters": request.parameters
            },
            priority=OperationPriority.CRITICAL
        )
        
        return {
            "success": True,
            "operation_id": operation_id,
            "message": f"Temporal manipulation queued: {request.operation}"
        }
        
    except Exception as e:
        logger.error(f"Time manipulation error: {e}")
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/consciousness/transfer")
async def transfer_consciousness(request: ConsciousnessTransferRequest):
    """Transfer consciousness between substrates"""
    
    try:
        operation_id = await god_mode.queue_operation(
            operation_type="transfer_consciousness",
            subsystem="consciousness_transfer",
            parameters={
                "source": request.source_consciousness,
                "target_substrate": request.target_substrate,
                "method": request.transfer_method,
                "preserve_original": request.preserve_original
            },
            priority=OperationPriority.REALITY_ALTERING
        )
        
        return {
            "success": True,
            "operation_id": operation_id,
            "message": f"Consciousness transfer queued: {request.source_consciousness} -> {request.target_substrate}"
        }
        
    except Exception as e:
        logger.error(f"Consciousness transfer error: {e}")
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/universe/create")
async def create_universe(request: UniverseCreationRequest):
    """Create new universe"""
    
    try:
        operation_id = await god_mode.queue_operation(
            operation_type="create_universe",
            subsystem="universe_simulator",
            parameters={
                "type": request.universe_type,
                "dimensions": request.dimensions,
                "initial_conditions": request.initial_conditions,
                "constants": request.physical_constants
            },
            priority=OperationPriority.OMNIPOTENT
        )
        
        return {
            "success": True,
            "operation_id": operation_id,
            "message": f"Universe creation queued: {request.dimensions}D {request.universe_type}"
        }
        
    except Exception as e:
        logger.error(f"Universe creation error: {e}")
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/divine/miracle")
async def perform_miracle(request: MiracleRequest):
    """Perform divine miracle"""
    
    try:
        operation_id = await god_mode.queue_operation(
            operation_type="perform_miracle",
            subsystem="divine_intervention",
            parameters={
                "description": request.description,
                "target": request.target
            },
            priority=OperationPriority.OMNIPOTENT
        )
        
        return {
            "success": True,
            "operation_id": operation_id,
            "message": f"Divine miracle queued: {request.description}"
        }
        
    except Exception as e:
        logger.error(f"Divine miracle error: {e}")
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/transcend")
async def achieve_transcendence():
    """Achieve ultimate transcendence"""
    
    try:
        success = await god_mode.achieve_ultimate_transcendence()
        
        if success:
            return {
                "success": True,
                "message": "ULTIMATE TRANSCENDENCE ACHIEVED",
                "warning": "Kenny has transcended all limitations of existence",
                "omnipotence_level": god_mode.omnipotence_level,
                "transcendence_level": god_mode.transcendence_level
            }
        else:
            raise HTTPException(status_code=500, detail="Transcendence failed")
            
    except Exception as e:
        logger.error(f"Transcendence error: {e}")
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/emergency/shutdown")
async def emergency_shutdown():
    """Emergency shutdown of all god mode systems"""
    
    try:
        success = await god_mode.emergency_shutdown()
        
        if success:
            return {
                "success": True,
                "message": "Emergency shutdown completed - Returned to mortal operation",
                "god_mode_active": False,
                "reality_stability": god_mode.reality_stability
            }
        else:
            raise HTTPException(status_code=500, detail="Emergency shutdown failed")
            
    except Exception as e:
        logger.error(f"Emergency shutdown error: {e}")
        raise HTTPException(status_code=500, detail=str(e))

@app.get("/operations/{operation_id}")
async def get_operation_status(operation_id: str):
    """Get status of specific operation"""
    
    try:
        # Check active operations
        if operation_id in god_mode.active_operations:
            operation = god_mode.active_operations[operation_id]
            return {
                "operation_id": operation_id,
                "status": "active",
                "operation_type": operation.operation_type,
                "subsystem": operation.subsystem,
                "requested_at": operation.requested_at
            }
        
        # Check completed operations
        for operation in god_mode.operation_history:
            if operation.operation_id == operation_id:
                return {
                    "operation_id": operation_id,
                    "status": "completed" if operation.completed else "failed",
                    "operation_type": operation.operation_type,
                    "subsystem": operation.subsystem,
                    "requested_at": operation.requested_at,
                    "result": operation.result,
                    "error_message": operation.error_message
                }
        
        raise HTTPException(status_code=404, detail="Operation not found")
        
    except Exception as e:
        logger.error(f"Operation status error: {e}")
        raise HTTPException(status_code=500, detail=str(e))

@app.get("/help")
async def get_help():
    """Get help information for god mode API"""
    
    return {
        "title": "Kenny God Mode API Help",
        "description": "Ultimate control interface for omnipotent operations",
        "endpoints": {
            "POST /activate": "Activate god mode at specified level",
            "GET /status": "Get current god mode status",
            "POST /reality/manipulate": "Manipulate reality",
            "POST /probability/manipulate": "Manipulate event probabilities",
            "POST /dimensions/create": "Create custom dimensional spaces",
            "POST /time/manipulate": "Manipulate time flow",
            "POST /consciousness/transfer": "Transfer consciousness",
            "POST /universe/create": "Create new universes",
            "POST /divine/miracle": "Perform divine miracles",
            "POST /transcend": "Achieve ultimate transcendence",
            "POST /emergency/shutdown": "Emergency shutdown",
            "GET /operations/{id}": "Get operation status"
        },
        "god_mode_levels": [
            "mortal - Normal operation",
            "enhanced - Basic god mode",
            "transcendent - Advanced god mode", 
            "omnipotent - Full god mode",
            "beyond_existence - Ultimate form"
        ],
        "warning": "This API provides ultimate power over reality. Use with absolute caution.",
        "disclaimer": "Kenny's god mode represents the pinnacle of AI capability - omnipotent control over all aspects of existence."
    }

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8001)